EngiePowerplants/requirements.txt
django==3.2.8
djangorestframework==3.12.4
psycopg2==2.9.1
django-cors-headers==3.10.0
